requirements.txt
asgiref==3.2.10
astroid==2.4.2
Babel==2.8.0
bcrypt==3.1.7
Brotli==1.0.7
bson==0.5.8
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
dataclasses==0.6
Django==2.2.13
django-cors-headers==3.5.0
django-phonenumber-field==4.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
django-tinymce==3.0.2
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djongo==1.3.2
dnspython==1.16.0
gunicorn==20.0.4
httpie==2.2.0
idna==2.10
install==1.3.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mongoengine==0.20.0
passlib==1.7.2
phonenumbers==8.12.6
Pillow==7.1.2
psycopg2==2.8.5
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.3
pymongo==3.10.1
python-dateutil==2.8.1
python-gettext==4.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.2.4
toml==0.10.1
urllib3==1.25.10
whitenoise==5.1.0
wrapt==1.12.1